Split playlist server helpers
